@@ -3,14 +3,15 @@
   "version": "0.0.39",
   "main": "index.js",
   "license": "MIT",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "npx ts-node -T src/index.ts",
     "dev": "nodemon src/index.ts",
-    "dev-ui": "vue-cli-service serve",
-    "build": "rm -rf ./dist/server && tsc --project ./tsconfig.json --outDir ./dist/server && vue-cli-service build",
+    "dev-ui": "vite",
+    "build": "rm -rf ./dist/server && tsc --project ./tsconfig.json --outDir ./dist/server && vite build",
     "build-server": "rm -rf ./dist/server && tsc --project ./tsconfig.json --outDir ./dist/server",
-    "build-ui": "vue-cli-service build",
+    "build-ui": "vite build",
     "run": "node dist/server/index.js"
   },
   "dependencies": {
@@ -22,52 +23,50 @@
     "@azure/openai": "^2.0.0",
     "@google/generative-ai": "^0.24.1",
     "@koa/cors": "^5.0.0",
-    "@larksuiteoapi/node-sdk": "^1.44.0",
-    "csv-parser": "^3.0.0",
+    "@larksuiteoapi/node-sdk": "^1.56.1",
+    "csv-parser": "^3.2.0",
     "dotenv": "17.2.1",
     "googleapis": "^144.0.0",
     "https-proxy-agent": "^7.0.6",
-    "koa": "^2.16.1",
+    "koa": "^2.16.3",
     "koa-body": "^6.0.1",
-    "koa-router": "^13.0.1",
+    "koa-router": "^13.1.1",
     "koa-static-server": "^1.5.2",
-    "nodemailer": "^6.9.16",
-    "ollama": "^0.5.15",
-    "openai": "^4.95.1",
-    "pg": "^8.14.1",
+    "nodemailer": "^6.10.1",
+    "ollama": "^0.5.18",
+    "openai": "^4.104.0",
+    "pg": "^8.17.1",
     "sharp": "0.34.3",
-    "winston": "^3.17.0"
+    "winston": "^3.19.0"
   },
   "compilerOptions": {
     "module": "NodeNext",
     "moduleResolution": "NodeNext"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
+    "@babel/core": "^7.28.6",
+    "@babel/eslint-parser": "^7.28.6",
     "@types/koa": "^2.15.0",
-    "@types/koa__cors": "^5.0.0",
-    "@types/koa-router": "^7.4.8",
-    "@types/pg": "^8.11.4",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "core-js": "^3.8.3",
+    "@types/koa-router": "^7.4.9",
+    "@types/koa__cors": "^5.0.1",
+    "@types/pg": "^8.16.0",
+    "@vitejs/plugin-vue2": "^2.3.4",
+    "@vitejs/plugin-vue2-jsx": "^1.1.1",
+    "@vueuse/core": "^8.9.4",
+    "dayjs": "^1.11.19",
     "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3",
-    "kui-icons": "^4.1.4",
-    "kui-vue": "^3.4.5",
-    "less": "^4.2.0",
+    "eslint-plugin-vue": "^8.7.1",
+    "hash-sum": "^2.0.0",
+    "kui-icons": "^4.1.6",
+    "kui-vue": "^3.6.8",
+    "less": "^4.5.1",
     "less-loader": "^11.1.4",
-    "nodemon": "^3.1.0",
-    "path-browserify": "^1.0.1",
+    "nodemon": "^3.1.11",
     "ts-node": "^10.9.2",
-    "typescript": "^5.7.2",
+    "typescript": "^5.9.3",
+    "vite": "npm:rolldown-vite@7.2.11",
     "vue": "^2.7.16",
-    "vue-clipboard2": "^0.3.3",
-    "vue-i18n": "^8.28.2",
     "vue-router": "^3.6.5",
-    "vue-template-compiler": "^2.6.14",
     "vuex": "^3.6.2"
   },
   "eslintConfig": {
@@ -96,5 +95,8 @@
       "sharp",
       "yorkie"
     ]
+  },
+  "overrides": {
+    "vite": "npm:rolldown-vite@7.2.11"
   }
 }
