@@ -1,247 +1,251 @@
 <template>
-  <div class="sys-tab-wraper">
-    <div class="sys-tab-box" ref="root">
-      <transition-group name="tab-fade" class="tab-inner-box" tag="div" @enter="enter" @beforeEnter="beforeEnter"
-        ref="tabbox" @afterEnter="afterEnter" @beforeLeave="beforeLeave" @leave="leave" @afterLeave="afterLeave">
-        <div :class="cls(view)" v-for="view in views" :key="view.path">
-          <Dropdown trigger="contextmenu" @click="e => handle(e, view)" :key="view.fullPath">
+  <div class="sys-tab-wrapper">
+    <div class="sys-tab-box" ref="rootRef">
+      <transition-group
+        class="tab-inner-box"
+        ref="tabBoxRef"
+        tag="div"
+        name="sys-tab"
+        :time="300"
+        @enter="animate.on.enter"
+        @beforeEnter="animate.on.beforeEnter"
+        @afterEnter="animate.on.afterEnter"
+        @beforeLeave="animate.on.beforeLeave"
+        @leave="animate.on.leave"
+        @afterLeave="animate.on.afterLeave"
+      >
+        <div :class="cls(view)" v-for="view in views" :key="view.key">
+          <Dropdown trigger="contextmenu" :key="view.fullPath">
             <div class="sys-tab-item-inner">
-              <router-link :to="{ path: view.path, query: view.query, fullPath: view.fullPath, params: view.params }">
-                <Icon class="sys-tab-icon" :type="view.meta.icon" v-if="view.meta.icon" />
-                <span class="sys-tab-title">{{ view.meta.title }}</span>
-                <Icon :type="Close" class="sys-tab-close" @click.prevent.stop="close(view)" :strokeWidth="40" />
+              <router-link
+                :to="{
+                  path: view.path,
+                  query: view.query,
+                  fullPath: view.fullPath,
+                  params: view.params,
+                }"
+              >
+                <Icon
+                  class="sys-tab-icon"
+                  :type="view.meta.icon"
+                  v-if="view.meta.icon"
+                />
+                <span class="sys-tab-title">{{ $t(view.meta.title) || "-" }}</span>
+                <Icon
+                  :type="Close"
+                  class="sys-tab-close"
+                  @click.prevent.stop="close(view)"
+                  :strokeWidth="40"
+                />
               </router-link>
             </div>
-            <Menu slot="content">
-              <MenuItem key="reload">重新加载</MenuItem>
-              <MenuItem key="close">关闭</MenuItem>
-              <MenuItem key="close-other">关闭其他</MenuItem>
-              <MenuItem key="close-left">关闭左侧</MenuItem>
-              <MenuItem key="close-right">关闭右侧</MenuItem>
-            </Menu>
+            <template #overlay>
+              <Menu @select="(e) => handle(e, view)">
+                <MenuItem key="reload">重新加载</MenuItem>
+                <MenuItem key="close">关闭</MenuItem>
+                <MenuItem key="close-other">关闭其他</MenuItem>
+                <MenuItem key="close-left">关闭左侧</MenuItem>
+                <MenuItem key="close-right">关闭右侧</MenuItem>
+              </Menu>
+            </template>
           </Dropdown>
         </div>
       </transition-group>
+      <!-- </draggable> -->
     </div>
-    <Dropdown trigger="hover" v-if="showDrop" @click="dropGo">
-      <Button :icon="ChevronDown" theme="light" size="small" class="sys-tab-show-list-btn"></Button>
-      <Menu slot="content">
-        <MenuItem :icon="view.meta.icon" v-for="view in views" :key="view.path">{{
-          view.meta.title }}</MenuItem>
-      </Menu>
+    <Dropdown trigger="hover" v-if="showDrop" placement="bottom" arrow>
+      <Button
+        :icon="ChevronDown"
+        theme="light"
+        size="small"
+        class="sys-tab-show-list-btn"
+      ></Button>
+      <template #overlay>
+        <Menu @select="dropGo">
+          <MenuItem
+            :icon="kui[view.meta.icon]"
+            v-for="view in views"
+            :key="view.fullPath"
+          >
+            {{ $t(view.meta.title) }}
+          </MenuItem>
+        </Menu>
+      </template>
     </Dropdown>
   </div>
 </template>
-<script>
-import { Close, Reload, Refresh, ChevronDown } from 'kui-icons'
-import { mapGetters } from 'vuex'
-// import { getTranstionHorProp } from '../utils/transition'
-export default {
-  name: 'Tab',
-  data() {
-    return {
-      Close, Reload, Refresh, ChevronDown,
-      showDrop: false
-    }
-  },
-  computed: {
-    ...mapGetters(['views']),
-    transProp() {
-      // return getTranstionHorProp('tab-fade')
-    },
-    current() {
-      return this.$route.path
-    },
-    currentIndex() {
-      return this.views.findIndex(v => v.path == this.current)
-    }
-  },
-  watch: {
-    '$route'() {
-      this.$store.commit('tabViews/addView', this.$route)
-      this.$nextTick(() => {
-        this.scrollToCenter()
-      })
-    }
-  },
-  beforeCreate() {
-    this.$store.commit('tabViews/addView', this.$route)
-  },
-  mounted() {
-    this.observe = new ResizeObserver((e) => {
-      // 调用 resize 方法，使图表自动适应新的容器尺寸
-      this.setDropShow(e[0].target, this._$('.tab-inner-box'))
-      this.scrollToCenter()
-    })
-    this.observe.observe(this._$('.sys-tab-box'))
-  },
-  methods: {
-    dropGo({ key }) {
-      let view = this.views.findLast(x => x.path == key)
-      // console.log(view)
-      this.go(view)
-    },
-    _$(clsName) {
-      return document.querySelector(clsName)
-    },
-    setDropShow(outter, inner) {
-      this.showDrop = outter.offsetWidth < inner.offsetWidth
-      // console.log(el.offsetWidth, this.$refs.tabbox.$el.offsetWidth)
-      // console.log(el.offsetWidth, el.scrollWidth, el.width)
-    },
-    scrollToCenter(animate = true) {
-      // console.log(e)
-      let root = this.$refs.root
-      let box = this.$refs.tabbox.$el
-      // console.log(root)
-      let childs = (box || {}).children || []
-      for (let m of childs) {
-        // console.log(m)
-        if (m.className.indexOf('actived') > -1) {
-          // 计算 span 元素相对于 div 元素的偏移
-          const offset = m.offsetLeft;
-          // 计算滚动距离使 span 元素垂直居中
-          const scrollDistance = offset - parseFloat((root.clientWidth / 2).toFixed(2)) + parseFloat((m.clientWidth / 2).toFixed(2));
-          // 滚动到计算出的位置
-          // console.log(scrollDistance, offset)
-          if (animate) {
-            root.scrollTo({ left: scrollDistance, behavior: 'smooth' });
-          } else {
-            root.scrollLeft = scrollDistance
-          }
-          break;
-        }
-      }
-    },
-    beforeEnter(el) {
-      el.style.overflow = 'hidden';
-      el.style.width = 0
-      el.style.opacity = 0.1
-    },
-    enter(el) {
-      if (el.scrollWidth !== 0) {
-        el.style.width = el.scrollWidth + 'px'
-        el.style.opacity = 1
-      } else {
-        el.style.width = ''
-        el.style.opacity = ''
-      }
-    },
-    afterEnter(el) {
-      el.style.width = ''
-      el.style.overflow = ''
-      el.style.opacity = ''
-    },
-    beforeLeave(el) {
-      el.style.width = el.scrollWidth + 'px'
-      el.style.opacity = 1
-    },
-    leave(el) {
-      if (el.scrollWidth !== 0) {
-        el.style.width = 0;
-        // el.style.paddingLeft = 0;
-        // el.style.paddingRight = 0;
-        // el.style.marginLeft = 0;
-        // el.style.marginRight = 0;
-        el.style.opacity = 0
-        // el.style.margin = 0
-        el.style.overflow = 'hidden';
-      }
-    },
-    afterLeave(el) {
-      el.style.width = '';
-      // el.style.paddingLeft = '';
-      // el.style.paddingRight = '';
-      // el.style.marginLeft = '';
-      // el.style.marginRight = '';
-      el.style.opacity = ''
-      el.style.overflow = ''
-    },
-    handle({ key }, view) {
-      let cur_index, select_index;
-      switch (key) {
-        case "reload":
+<script setup>
+import { ref, computed, watch, onMounted, onBeforeMount, nextTick,inject } from "vue";
+import { getCurrentInstance } from "vue";
+import * as kui from "kui-icons";
+import { Close, ChevronDown, Loading } from "kui-icons";
+import id from "hash-sum";
+import { getTransitionHorProp } from "@/utils/transition";
+const animate = getTransitionHorProp("tab-fade");
+const { proxy } = getCurrentInstance();
+const $t = inject("$t");
+const showDrop = ref(false);
+const observe = ref(null);
+const rootRef = ref(null);
+const tabBoxRef = ref(null);
 
-          this.reload(view)
-          break;
-        case "close":
-          this.close(view)
-        case "close-other":
-          this.$store.commit('tabViews/closeOtherView', view)
-          if (this.current != view.path) {
-            this.go(view)
-          }
-          break;
-        case "close-right":
-          // return;
-          cur_index = this.views.findIndex(x => x.path == this.current)
-          select_index = this.views.indexOf(view)
-          // console.log(cur_index, select_index, this.views)
-          if (this.current != view.path && cur_index > select_index) {
-            this.go(view)
-          }
-          this.$store.commit('tabViews/closeRightView', view)
-          break;
-        case "close-left":
-          cur_index = this.views.findIndex(x => x.path == this.current)
-          select_index = this.views.indexOf(view)
-          if (this.current != view.path && cur_index < select_index) {
-            this.go(view)
-          }
-          this.$store.commit('tabViews/closeLeftView', view)
-          break;
-        default:
-          break
-      }
-    },
-    reload(view) {
-      this.go(view)
-      this.$store.commit('tabViews/reloadView', view)
-    },
-    close(view) {
-      let views = this.views
-      if (views.length == 1) return
-      if (view.path == this.current) {
-        let index = views.findIndex(x => x.path == view.path)
+const views = computed(() => proxy.$store.state.tabViews.views); //getters.views);
+const current = computed(() => proxy.$route.fullPath);
+const currentIndex = computed(() =>
+  views.value.findIndex((v) => v.fullPath == current.value)
+);
 
-        if (index == 0) {
-          index = 1
-        } else if (index == views.length - 1) {
-          index = views.length - 2
-        } else {
-          index += 1
-        }
-        let item = views[index]
-        this.$store.commit('tabViews/closeView', view)
-        this.go(item)
-      } else {
-        this.$store.commit('tabViews/closeView', view)
-      }
-    },
-    go(item) {
-      this.$router.push({ path: item.path, query: item.query, fullPath: item.fullPath, params: item.params })
-    },
-    cls(item) {
-      let index = this.views.indexOf(item)
-      return ['sys-tab-item', {
-        'sys-tab-item-actived': item.path == this.current,
-        'sys-tab-item-first': index == 0,
-        'sys-tab-item-prev': index == this.currentIndex - 1,
-        'sys-tab-item-next': index == this.currentIndex + 1,
-      }]
-    },
+watch(
+  () => proxy.$route,
+  (newRoute) => {
+    proxy.$store.commit("tabViews/addView", newRoute);
+    nextTick(() => {
+      scrollToCenter();
+      setDropShow(_$(".sys-tab-box"), tabBoxRef.value?.$el);
+    });
   }
-}
+);
+
+onBeforeMount(() => {
+  proxy.$store.commit("tabViews/addView", proxy.$route);
+});
+
+onMounted(() => {
+  observe.value = new ResizeObserver((e) => {
+    setDropShow(e[0].target, _$(".tab-inner-box"));
+    scrollToCenter();
+  });
+  observe.value.observe(_$(".sys-tab-box"));
+});
+
+const dropGo = ({ key }) => {
+  let view = views.value.findLast((x) => x.fullPath == key);
+  go(view);
+};
+
+const _$ = (clsName) => {
+  return document.querySelector(clsName);
+};
+
+const setDropShow = (outer, inner) => {
+  if (!inner) return;
+  showDrop.value = outer.offsetWidth < inner.offsetWidth;
+};
+
+const scrollToCenter = (animate = true) => {
+  let box = tabBoxRef.value?.$el;
+  let children = box?.children || [];
+  for (let m of children) {
+    if (m.className.indexOf("active") > -1) {
+      const offset = m.offsetLeft;
+      const scrollDistance =
+        offset -
+        parseFloat((rootRef.value.clientWidth / 2).toFixed(2)) +
+        parseFloat((m.clientWidth / 2).toFixed(2));
+      if (animate) {
+        rootRef.value.scrollTo({ left: scrollDistance, behavior: "smooth" });
+      } else {
+        rootRef.value.scrollLeft = scrollDistance;
+      }
+      break;
+    }
+  }
+};
+
+const handle = ({ key }, view) => {
+  let cur_index, select_index;
+  switch (key) {
+    case "reload":
+      reload(view);
+      break;
+    case "close":
+      close(view);
+      break;
+    case "close-other":
+      proxy.$store.commit("tabViews/closeOtherView", view);
+      if (current.value != view.fullPath) {
+        go(view);
+      }
+      break;
+    case "close-right":
+      cur_index = views.value.findIndex((x) => x.fullPath == current.value);
+      select_index = views.value.indexOf(view);
+      if (current.value != view.fullPath && cur_index > select_index) {
+        go(view);
+      }
+      proxy.$store.commit("tabViews/closeRightView", view);
+      break;
+    case "close-left":
+      cur_index = views.value.findIndex((x) => x.fullPath == current.value);
+      select_index = views.value.indexOf(view);
+      if (current.value != view.fullPath && cur_index < select_index) {
+        go(view);
+      }
+      proxy.$store.commit("tabViews/closeLeftView", view);
+      break;
+    default:
+      break;
+  }
+};
+
+const reload = (view) => {
+  let currentId = id(proxy.$route.fullPath);
+  if (currentId != view.key) {
+    proxy.$store.commit("tabViews/reloadSelectView", view);
+    return;
+  }
+  proxy.$store.commit("tabViews/reloadView", view);
+};
+
+const close = (view) => {
+  let viewsArray = views.value;
+  if (viewsArray.length == 1) return;
+  if (view.fullPath == current.value) {
+    let index = viewsArray.findIndex((x) => x.fullPath == view.fullPath);
+
+    if (index == 0) {
+      index = 1;
+    } else if (index == viewsArray.length - 1) {
+      index = viewsArray.length - 2;
+    } else {
+      index += 1;
+    }
+    let item = viewsArray[index];
+    proxy.$store.commit("tabViews/closeView", view);
+    go(item);
+  } else {
+    proxy.$store.commit("tabViews/closeView", view);
+  }
+};
+
+const go = (item) => {
+  proxy.$router.push({
+    path: item.path,
+    query: item.query,
+    fullPath: item.fullPath,
+    params: item.params,
+  });
+};
+
+const cls = (item) => {
+  let index = views.value.indexOf(item);
+  return [
+    "sys-tab-item",
+    {
+      "sys-tab-item-active": item.fullPath == current.value,
+      "sys-tab-item-first": index == 0,
+      "sys-tab-item-prev": index == currentIndex.value - 1,
+      "sys-tab-item-next": index == currentIndex.value + 1,
+    },
+  ];
+};
 </script>
 <style lang="less">
-.sys-tab-wraper {
+.sys-tab-wrapper {
   display: flex;
   align-items: center;
 
   .sys-tab-show-list-btn.k-btn-sm {
-    border-radius: 10px;
-    width: 27px;
-    height: 27px;
     margin: 0 8px;
   }
 
@@ -250,7 +254,8 @@ export default {
     flex-wrap: nowrap;
     overflow: auto;
     gap: 2px;
-    padding: 8px 10px 0;
+    padding: 8px 0 0;
+    flex: 1;
 
     &::-webkit-scrollbar {
       height: 0;
@@ -259,6 +264,9 @@ export default {
     .tab-inner-box {
       display: flex;
       flex-wrap: nowrap;
+      margin-left: 8px;
+      height: 34px;
+      transition: all 0.3s ease-in-out;
     }
 
     .sys-tab-item {
@@ -267,60 +275,69 @@ export default {
       cursor: pointer;
       height: 26px;
       line-height: 26px;
-      // color: #495060;
       font-size: 12px;
-
+      margin: 0 4px;
       &::before,
       &::after {
-        content: '';
+        content: "";
         position: absolute;
-        width: 2px;
-        background: var(--kui-color-main-80);
-        height: 13px;
-        z-index: 10;
-        border-radius: 3px;
+        width: 15px;
+        height: 17px;
+        background: none;
+        bottom: 8px;
+        z-index: 0;
       }
 
       &::before {
-        left: -1px;
-        top: 7px;
+        box-shadow: 12px 12px 0px 9px var(--kui-color-bg-layout);
+        border-radius: 0 0 12px 0;
+        left: -15px;
+        bottom: 2px;
+        height: 16px;
+        top: auto;
       }
 
       &::after {
-        right: -1px;
-        top: 7px;
-      }
-
-      &:hover {
-        z-index: 99;
-
-        &::before,
-        &::after {
-          transition: background-color .3s;
-          background-color: var(--kui-color-nav-header);
-        }
-
-        .sys-tab-item-inner {
-          // transition: background-color .3s;
-          background: var(--kui-color-main-80)
-        }
+        border-radius: 0 0 0 12px;
+        box-shadow: -12px 12px 0px 9px var(--kui-color-bg-layout);
+        right: -15px;
+        top: 15px;
       }
 
       .sys-tab-item-inner {
         border-radius: 8px;
-        margin: 0 3px;
-        // transition: background-color .3s ease-in;
-        // background-color: #cccccc20;
+        z-index: 1;
+        position: relative;
+        &::before,
+        &::after {
+          content: "";
+          position: absolute;
+          width: 2px;
+          height: 14px;
+          background: var(--kui-color-item-hover);
+          top: 6px;
+        }
+        &::before {
+          left: -5px;
+        }
+        &::after {
+          right: -5px;
+        }
 
+        &:hover {
+          background: var(--kui-color-item-hover);
+        }
+        &:active {
+          background: var(--kui-color-item-active);
+        }
         a {
           padding: 0 8px;
-          color: var(--kui-color-font);
+          color: var(--kui-color-text);
           display: flex;
           text-decoration: none;
           align-items: center;
         }
       }
-
 
       .sys-tab-title {
         max-width: 120px;
@@ -342,140 +359,59 @@ export default {
         z-index: 10;
 
         &:hover {
-          background-color: var(--kui-color-gray-60);
+          background-color: var(--kui-color-bg-component-hover);
+        }
+        &:active {
+          background-color: var(--kui-color-bg-component-active);
         }
       }
     }
-
-    .sys-tab-item-first:not(.sys-tab-item-actived) {
-      &::before {
-        display: none;
-      }
-    }
-
+    .sys-tab-item-first,
     .sys-tab-item-next {
-      &::before {
-        display: none;
-        // background-color: var(--kui-color-border);
+      .sys-tab-item-inner::before {
+        background-color: transparent;
       }
     }
-
-
-
     .sys-tab-item-prev {
-      &::after {
-        display: none;
-        background-color: var(--kui-color-border);
+      .sys-tab-item-inner::after {
+        background-color: transparent;
       }
     }
 
-    .sys-tab-item-actived {
-      // background-color: #fff;
-      // color: #000;
-      // border-color: #42b983;
-      // height: 40px;
+    .sys-tab-item-active {
       vertical-align: top;
       align-self: flex-start;
       justify-self: baseline;
       border-radius: 10px 10px 0 0;
       padding-bottom: 8px;
       position: relative;
-      transition: none;
       z-index: 100;
-
+      background-color: var(--kui-color-bg);
+      &::before {
+        box-shadow: 12px 12px 0px 9px var(--kui-color-bg);
+      }
+      &::after {
+        box-shadow: -12px 12px 0px 9px var(--kui-color-bg);
+      }
       .sys-tab-item-inner {
-        background: var(--kui-color-back);
+        background: var(--kui-color-bg);
         position: relative;
         border-radius: 8px 8px 0 0;
         z-index: 1;
-
+        &:hover {
+          background: var(--kui-color-bg);
+        }
+        &::before,
         &::after {
-          content: '';
-          position: absolute;
-          width: 100%;
-          bottom: -8px;
-          left: 0;
-          height: 8px;
-          background-color: var(--kui-color-back);
-        }
-      }
-
-      &:hover {
-        .sys-tab-item-inner {
-          background: var(--kui-color-back);
-        }
-
-        &::after,
-        &::before {
-          transition: none;
           background-color: transparent;
-          display: block;
         }
       }
-
-      &::before,
-      &::after {
-        content: '';
-        transition: none;
-        position: absolute;
-        width: 15px;
-        height: 17px;
-        background: none;
-        bottom: 8px;
-        z-index: 0;
-      }
-
-      &::before {
-        box-shadow: 12px 12px 0px 9px var(--kui-color-back);
-        border-radius: 0 0 12px 0;
-        left: -12px;
-        bottom: 2px;
-        height: 16px;
-        top: auto;
-      }
-
-      &::after {
-        border-radius: 0 0 0 12px;
-        box-shadow: -12px 12px 0px 9px var(--kui-color-back);
-        right: -12px;
-        top: 15px;
-      }
-    }
-
-
-  }
-
-  @keyframes fadein {
-    0% {
-      opacity: 0;
-      width: 0;
-    }
-
-    100% {
-      opacity: 1;
     }
   }
-
-  @-webkit-keyframes fadein {
-    0% {
-      width: 0;
-      opacity: 0;
-    }
-
-    100% {
-      opacity: 1;
-    }
-  }
-
-  // .tab-fade-enter-active {
-  //   animation: fadein 0.3s ease-in;
-  //   animation-fill-mode: both;
-  // }
 
   .tab-fade-leave-active {
-    transition: width .3s;
+    transition: width 0.3s;
     animation: fadeout 0.3s ease-in-out;
-    // animation-fill-mode: both;
   }
 
   @keyframes fadeout {
@@ -499,5 +435,39 @@ export default {
       opacity: 0;
     }
   }
+}
+
+.sys-tab-enter,
+.sys-tab-leave-to {
+  transition: all 0.3s ease-in-out;
+  opacity: 0;
+  .sys-tab-item-inner {
+    transition: all 0.3s ease-in-out;
+  }
+}
+.sys-tab-enter-active,
+.sys-tab-leave-active {
+  transition: all 0.3s ease-in-out;
+  .sys-tab-item-inner {
+    transition: all 0.3s ease-in-out;
+  }
+}
+/* 可选：显式定义 enter-to（其实默认 opacity:1，可省略） */
+.sys-tab-enter-to,
+.sys-tab-leave {
+  opacity: 1;
+
+  .sys-tab-item-inner {
+    transition: all 0.3s ease-in-out;
+  }
+}
+
+.k-dropdown-menu {
+  max-height: calc(100vh - 100px);
+  overflow: auto;
+}
+
+.k-dropdown-menu-item {
+  height: 32px !important;
 }
 </style>
